Display delete instructions in the payload
This refactors the Byzcoin payload component and add the support
of delete instructions.

Fixes #18
